+++ src/Svelte/Common/Pp.svelte
@@ -1,0 +1,9 @@
+<script>
+    import Value from './Value.svelte';
+
+    export let pp = 0;
+    export let zero = '-';
+</script>
+
+<span class="scoreTop ppValue"><Value value="{pp}" {zero} /></span>
+<span class="scoreTop ppLabel">pp</span>